+++ spine-ios/Sources/Spine/Metal/SpineObjects.swift
@@ -5,15 +5,15 @@ import MetalKit
 ///
 /// Persistent Objects
 /// https://developer.apple.com/library/archive/documentation/3DDrawing/Conceptual/MTLBestPracticesGuide/PersistentObjects.html#//apple_ref/doc/uid/TP40016642-CH3-SW1
-final class SpineObjects {
+internal final class SpineObjects {
     
     static let shared = SpineObjects()
     
-    lazy var device: MTLDevice = {
+    internal lazy var device: MTLDevice = {
         MTLCreateSystemDefaultDevice()!
     }()
     
-    lazy var commandQueue: MTLCommandQueue = {
+    internal lazy var commandQueue: MTLCommandQueue = {
         device.makeCommandQueue()!
     }()
 }